Refactor
Big changes include:
removeIdentifiers() method - when passed a statement, it will go
through each element in the statement and replace each identifier
with its actual value. This helps reduce large if statements. This
function modifies the statement itself and won't make a copy.

evaluate() methods - Edited to make use of removeIdentifiers() as
well as had its functionality slightly changed to change the
statement passed to it in place. Doing this also helps us cut down
on code in other functions.